Calculate the distance between colours

There are many ways to measure the distance between colours. farver provides 5 different algorithms, ranging from simple euclidean distance in RGB space, to different perceptual measures such as CIE2000.

compare_colour(from, to = NULL, from_space, to_space = from_space,
  method = "euclidean", white_from = "D65", white_to = white_from)

Arguments

from, to
Numeric matrices with colours to compare - the format is the same as that for convert_colour(). If to is not set from will be compared with itself and only the upper triangle will get calculated

from_space, to_space
The colour space of from and to respectively. to_space defaults to be the same as from_space.

method
The method to use for comparison. Either 'euclidean', 'cie1976', 'cie94', 'cie2000', or 'cmc'

white_from, white_to
The white reference of the from and to colour space. Will only have an effect for relative colour spaces such as Lab and luv. Any value accepted by as_white_ref() allowed.

Value

A numeric matrix with the same number of rows as colours in from and the same number of columns as colours in to. If to is not given, only the upper triangle will be returned.

Handling of non-finite and out of bounds values

NA, NaN, -Inf, and Inf are treated as invalid input and will result in NA values for the colour. If a given colourspace has finite bounds in some of their channels, the input will be capped before conversion, and the output will be capped before returning, so that both input and output colours are valid colours in their respective space. This means that converting back and forth between two colourspaces may result in a change in the colour if the gamut of one of the spaces is less than the other.

Examples

r <- decode_colour(rainbow(10))
h <- decode_colour(heat.colors(15))

# Compare two sets of colours
compare_colour(r, h, 'rgb', method = 'cie2000')

#>           [,1]       [,2]      [,3]     [,4]     [,5]     [,6]      [,7]
#>  [1,]  0.00000  0.9533762  3.000824  6.57770 11.40975 17.52601 24.723573
#>  [2,] 29.50083 28.5558177 26.521813 22.95539 18.11996 11.98805  4.775709
#>  [3,] 72.33606 71.3575538 69.242820 65.50814 60.39647 53.85364 46.123120
#>  [4,] 85.84698 84.7949077 82.516916 78.48359 72.95656 65.91025 57.700879
#>  [5,] 85.92110 84.8858340 82.646223 78.68939 73.29065 66.45879 58.585358
#>  [6,] 70.95853 70.2695051 68.802658 66.28933 63.01735 59.12704 54.973704
#>  [7,] 50.38091 50.6204430 51.184821 52.32043 54.12124 56.78372 60.394557
#>  [8,] 52.15698 52.7007613 53.923828 56.23309 59.66383 64.36086 69.987595
#>  [9,] 44.47475 44.9165073 45.932292 47.90510 50.91157 55.17495 60.707919
#> [10,] 30.89123 31.3674680 32.487084 34.71591 38.15422 42.98987 49.095286
#>           [,8]     [,9]    [,10]    [,11]      [,12]      [,13]    [,14]
#>  [1,] 32.67872 40.97546 49.52951 57.25740  64.299474  63.895520 60.57729
#>  [2,]  3.16786 11.38266 19.71938 27.08742  33.646908  33.488301 32.47947
#>  [3,] 37.67285 29.15780 20.92325 14.21750   9.325976   9.059491 10.49557
#>  [4,] 48.97226 40.53408 32.80674 26.89319  22.660408  22.341120 21.56970
#>  [5,] 50.32473 42.44960 35.33751 29.96925  26.166094  25.707067 23.60073
#>  [6,] 50.98220 47.53531 44.77211 42.98433  41.972721  41.436516 37.40409
#>  [7,] 64.89992 70.06575 75.76079 81.12373  86.129806  85.536298 80.31417
#>  [8,] 76.18219 82.84641 89.96539 96.53168 102.567205 101.830677 95.66170
#>  [9,] 67.30314 74.52107 82.12052 88.99246  95.209202  94.367152 87.31037
#> [10,] 56.09295 63.45764 70.99828 77.74381  83.865025  83.098194 76.75948
#>          [,15]
#>  [1,] 52.36649
#>  [2,] 32.00440
#>  [3,] 21.49275
#>  [4,] 26.48901
#>  [5,] 25.76501
#>  [6,] 28.94258
#>  [7,] 64.93634
#>  [8,] 78.99746
#>  [9,] 67.87546
#> [10,] 59.84362

# Compare a set of colours with itself
compare_colour(r, from_space = 'rgb', method = 'cmc')

#>       [,1]     [,2]      [,3]      [,4]      [,5]     [,6]      [,7]     [,8]
#>  [1,]    0 10.52597 18.003563 15.958871 15.623504 23.06674  5.030710 11.74396
#>  [2,]    0  0.00000  9.732855 12.581267  7.948242 12.62015  9.568638 21.39777
#>  [3,]    0  0.00000  0.000000  5.729119  1.877060 14.58716 16.089339 22.78038
#>  [4,]    0  0.00000  0.000000  0.000000  5.253523 18.88463 16.297394 19.60278
#>  [5,]    0  0.00000  0.000000  0.000000  0.000000 14.46503 14.540410 21.56755
#>  [6,]    0  0.00000  0.000000  0.000000  0.000000  0.00000 21.444871 37.66022
#>  [7,]    0  0.00000  0.000000  0.000000  0.000000  0.00000  0.000000 14.69552
#>  [8,]    0  0.00000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#>  [9,]    0  0.00000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> [10,]    0  0.00000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#>            [,9]     [,10]
#>  [1,]  3.885606  5.174003
#>  [2,] 13.496209  6.515862
#>  [3,] 15.389917 13.626292
#>  [4,] 12.801400 14.321552
#>  [5,] 13.988362 12.037620
#>  [6,] 29.707650 18.989681
#>  [7,]  8.957990  3.701201
#>  [8,] 11.098901 17.436529
#>  [9,]  0.000000  8.718964
#> [10,]  0.000000  0.000000

# Compare colours from different colour spaces
h_luv <- convert_colour(h, 'rgb', 'luv')
compare_colour(r, h_luv, 'rgb', 'luv')

#>            [,1]      [,2]      [,3]      [,4]      [,5]     [,6]      [,7]
#>  [1,] 267.09465 264.12635 258.00957 248.10288 236.11058 222.8477 209.40537
#>  [2,] 202.95752 201.92552 199.96948 197.31603 195.02436 193.7811 194.05775
#>  [3,] 170.65985 171.89180 174.66702 179.83088 187.17551 196.6858 207.84034
#>  [4,]  80.00917  83.81549  91.73431 104.78824 121.01489 139.6057 159.31600
#>  [5,]  96.07824  99.60816 107.02674 119.43640 135.09965 153.2815 172.76260
#>  [6,]  96.07824  99.60816 107.02674 119.43640 135.09965 153.2815 172.76260
#>  [7,]  90.37049  87.74164  82.60672  75.32912  68.94652  66.2811  69.47742
#>  [8,] 175.03655 171.25923 163.42306 150.57539 134.74959 116.8889  98.44753
#>  [9,] 231.00024 227.70919 220.90831 209.83688 196.33522 181.2693 165.85516
#> [10,] 267.09465 264.12635 258.00957 248.10288 236.11058 222.8477 209.40537
#>            [,8]      [,9]     [,10]     [,11]     [,12]    [,13]    [,14]
#>  [1,] 196.59807 184.92467 174.21224 165.42351 158.04933 157.9709 157.3363
#>  [2,] 195.93552 199.21449 203.78168 208.92152 214.53825 214.6376 215.4115
#>  [3,] 219.98325 232.51965 245.54797 257.70046 269.38309 269.5585 270.9375
#>  [4,] 179.16606 198.50882 217.76631 235.17863 251.54893 251.7780 253.5931
#>  [5,] 192.54547 211.95095 231.37765 249.02371 265.67663 265.9065 267.7329
#>  [6,] 192.54547 211.95095 231.37765 249.02371 265.67663 265.9065 267.7329
#>  [7,]  78.24211  90.67485 105.57722 120.48879 135.36986 135.5721 137.1915
#>  [8,]  80.77083  65.16866  52.76093  46.47817  46.78045  46.8116  47.1220
#>  [9,] 151.04993 137.50123 125.12137 115.15537 107.13914 107.0558 106.3869
#> [10,] 196.59807 184.92467 174.21224 165.42351 158.04933 157.9709 157.3363
#>          [,15]
#>  [1,] 155.9575
#>  [2,] 216.9021
#>  [3,] 273.6799
#>  [4,] 257.2890
#>  [5,] 271.4785
#>  [6,] 271.4785
#>  [7,] 140.5874
#>  [8,]  48.1107
#>  [9,] 104.9653
#> [10,] 155.9575