Modify colour space channels in hex-encoded colour strings

This set of functions allows you to modify colours as given by strings, whithout first decoding them. For large vectors of colour values this should provide a considerable speedup.

set_channel(colour, channel, value, space = "rgb", white = "D65")

add_to_channel(colour, channel, value, space = "rgb", white = "D65")

multiply_channel(colour, channel, value, space = "rgb", white = "D65")

raise_channel(colour, channel, value, space = "rgb", white = "D65")

cap_channel(colour, channel, value, space = "rgb", white = "D65")

get_channel(colour, channel, space = "rgb", white = "D65")

Arguments

colour
A character string giving colours, either as hexadecimal strings or accepted colour names.

channel
The channel to modify or extract as a single letter, or 'alpha' for the alpha channel.

value
The value to modify with

space
The colour space the channel pertains to. Allowed values are: "cmy", "cmyk", "hsl", "hsb", "hsv", "lab" (CIE L*ab), "hunterlab" (Hunter Lab), "lch" (CIE Lch(ab) / polarLAB), "luv", "rgb" (sRGB), "xyz", "yxy" (CIE xyY), or "hcl" (CIE Lch(uv) / polarLuv)

white
The white reference of the channel colour space. Will only have an effect for relative colour spaces such as Lab and luv. Any value accepted by as_white_ref() allowed.

Value

A character vector of the same length as colour (or a numeric vector in the case of get_channel())

See also