as_white_ref {farver}	R Documentation

Convert value to a tristimulus values normalised to Y=100

Description

This function can take either the name of a standardised illuminants, x and y chromaticity coordinates or X, Y, and Z tristimulus values and converts it to tristimulus values normalised to Y=100. All Illuminant series A-F are supported and can be queried both on the CIE 1931 2° and CIE 1964 10° chromaticity coordinates.

Usage

as_white_ref(x, fow = 2)

Arguments

x
A string giving the name of the standardized illuminant or a 2 (chromaticity) or 3 (trsitimulus) length numeric vector.

fow
The field-of-view for the illuminant - either 2 or 10

Value

A 3-length vector with tristimulus values

Examples

# Using names
as_white_ref('D65')

# Using chromaticity values
as_white_ref(c(0.3, 0.4))

[Package farver version 2.1.0 Index]