Calculate the distance between colours

There are many ways to measure the distance between colours. farver provides 5 different algorithms, ranging from simple euclidean distance in RGB space, to different perceptual measures such as CIE2000.

compare_colour(from, to = NULL, from_space, to_space = from_space,
  method = "euclidean", white_from = "D65", white_to = white_from)

Arguments

from, to
Numeric matrices with colours to compare - the format is the same as that for convert_colour(). If to is not set from will be compared with itself and only the upper triangle will get calculated

from_space, to_space
The colour space of from and to respectively. to_space defaults to be the same as from_space.

method
The method to use for comparison. Either 'euclidean', 'cie1976', 'cie94', 'cie2000', or 'cmc'

white_from, white_to
The white reference of the from and to colour space. Will only have an effect for relative colour spaces such as Lab and luv. Any value accepted by as_white_ref() allowed.

Value

A numeric matrix with the same number of rows as colours in from and the same number of columns as colours in to. If to is not given, only the upper triangle will be returned.

Handling of non-finite and out of bounds values

NA, NaN, -Inf, and Inf are treated as invalid input and will result in NA values for the colour. If a given colourspace has finite bounds in some of their channels, the input will be capped before conversion, and the output will be capped before returning, so that both input and output colours are valid colours in their respective space. This means that converting back and forth between two colourspaces may result in a change in the colour if the gamut of one of the spaces is less than the other.

Examples

r <- decode_colour(rainbow(10))
h <- decode_colour(heat.colors(15))

# Compare two sets of colours
compare_colour(r, h, 'rgb', method = 'cie2000')

#>            [,1]     [,2]     [,3]     [,4]     [,5]      [,6]      [,7]
#>  [1,] 61.290565 59.83918 58.01212 55.75910 52.89779  40.67956  42.56588
#>  [2,] 32.656753 32.31764 32.01258 31.83175 31.93973  60.95501  61.89683
#>  [3,]  9.808456 11.31743 13.62816 16.71313 20.73983  98.09925  99.04616
#>  [4,] 21.516399 21.71896 22.42260 23.77938 26.02469 109.26968 110.21768
#>  [5,] 23.847781 23.44082 23.41026 23.96920 25.42038 107.18610 107.89415
#>  [6,] 38.229289 36.56742 34.56043 32.19957 29.42029  61.12095  57.81236
#>  [7,] 81.483235 79.07903 75.90901 71.76359 66.05914  32.20222  31.26769
#>  [8,] 97.006499 94.25759 90.71934 86.20707 80.16633  31.07699  31.44609
#>  [9,] 88.851113 85.69913 81.62389 76.38516 69.26957  10.35214   9.83255
#> [10,] 78.132842 75.32853 71.73190 67.15550 61.03031  12.54220  15.43917
#>            [,8]      [,9]     [,10]    [,11]     [,12]    [,13]    [,14]
#>  [1,]  42.53779  42.52180  42.55738 17.52601 24.723573 32.67872 40.97546
#>  [2,]  61.37806  60.47016  59.23008 11.98805  4.775709  3.16786 11.38266
#>  [3,]  98.08959  96.37053  93.93866 53.85364 46.123120 37.67285 29.15780
#>  [4,] 109.14207 107.20458 104.45616 65.91025 57.700879 48.97226 40.53408
#>  [5,] 106.83887 104.93741 102.23857 66.45879 58.585358 50.32473 42.44960
#>  [6,]  57.45840  56.86591  56.10718 59.12704 54.973704 50.98220 47.53531
#>  [7,]  31.42684  31.76641  32.33987 56.78372 60.394557 64.89992 70.06575
#>  [8,]  32.56564  34.53167  37.06742 64.36086 69.987595 76.18219 82.84641
#>  [9,]  10.70299  12.31163  14.58932 55.17495 60.707919 67.30314 74.52107
#> [10,]  15.59627  16.03766  16.92732 42.98987 49.095286 56.09295 63.45764
#>          [,15]
#>  [1,] 49.52951
#>  [2,] 19.71938
#>  [3,] 20.92325
#>  [4,] 32.80674
#>  [5,] 35.33751
#>  [6,] 44.77211
#>  [7,] 75.76079
#>  [8,] 89.96539
#>  [9,] 82.12052
#> [10,] 70.99828

# Compare a set of colours with itself
compare_colour(r, from_space = 'rgb', method = 'cmc')

#>       [,1]     [,2]      [,3]      [,4]      [,5]     [,6]      [,7]     [,8]
#>  [1,]    0 10.52597 18.003563 15.958871 15.623504 23.06674  5.030710 11.74396
#>  [2,]    0  0.00000  9.732855 12.581267  7.948242 12.62015  9.568638 21.39777
#>  [3,]    0  0.00000  0.000000  5.729119  1.877060 14.58716 16.089339 22.78038
#>  [4,]    0  0.00000  0.000000  0.000000  5.253523 18.88463 16.297394 19.60278
#>  [5,]    0  0.00000  0.000000  0.000000  0.000000 14.46503 14.540410 21.56755
#>  [6,]    0  0.00000  0.000000  0.000000  0.000000  0.00000 21.444871 37.66022
#>  [7,]    0  0.00000  0.000000  0.000000  0.000000  0.00000  0.000000 14.69552
#>  [8,]    0  0.00000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#>  [9,]    0  0.00000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#> [10,]    0  0.00000  0.000000  0.000000  0.000000  0.00000  0.000000  0.00000
#>            [,9]     [,10]
#>  [1,]  3.885606  5.174003
#>  [2,] 13.496209  6.515862
#>  [3,] 15.389917 13.626292
#>  [4,] 12.801400 14.321552
#>  [5,] 13.988362 12.037620
#>  [6,] 29.707650 18.989681
#>  [7,]  8.957990  3.701201
#>  [8,] 11.098901 17.436529
#>  [9,]  0.000000  8.718964
#> [10,]  0.000000  0.000000

# Compare colours from different colour spaces
h_luv <- convert_colour(h, 'rgb', 'luv')
compare_colour(r, h_luv, 'rgb', 'luv')

#>            [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#>  [1,] 221.42391 223.33260 222.10259 220.19485 217.94653 258.49652 252.61449
#>  [2,] 211.00494 208.71472 207.34795 204.88400 201.48888 191.50105 186.61530
#>  [3,] 222.60055 217.83061 216.81894 214.76743 211.67223 160.56650 158.64023
#>  [4,] 163.79160 157.88736 157.85141 157.37370 156.25322  81.15444  85.73625
#>  [5,] 172.21280 166.80877 167.20229 167.48420 167.39886 102.91513 108.70157
#>  [6,] 172.21280 166.80877 167.20229 167.48420 167.39886 102.91513 108.70157
#>  [7,]  54.32044  54.11822  55.50671  57.85567  60.93683  97.60800  97.94306
#>  [8,]  91.25125  97.18253  97.30218  97.99838  99.54864 175.56301 172.20740
#>  [9,] 176.21041 178.79094 177.65443 175.96093 174.06895 223.64689 217.87935
#> [10,] 221.42391 223.33260 222.10259 220.19485 217.94653 258.49652 252.61449
#>            [,8]      [,9]     [,10]     [,11]     [,12]     [,13]    [,14]
#>  [1,] 243.63170 231.07525 216.36111 200.73884 185.14279 175.73159 167.7826
#>  [2,] 181.03938 175.42619 170.58260 167.88722 167.58776 174.99006 183.4128
#>  [3,] 158.83883 162.52032 169.12935 178.93751 190.96264 207.27824 223.3232
#>  [4,]  94.53898 108.41226 125.43164 144.63089 164.85250 183.56105 201.9296
#>  [5,] 117.94448 131.41793 147.80578 166.19842 185.69747 202.50626 219.4242
#>  [6,] 117.94448 131.41793 147.80578 166.19842 185.69747 202.50626 219.4242
#>  [7,]  96.33300  93.16703  91.36879  92.30317  97.28179 100.28388 106.9846
#>  [8,] 165.01373 153.11953 138.72975 122.84691 107.17590  90.10055  74.9450
#>  [9,] 208.61852 195.20560 179.21379 161.83912 144.14249 131.86251 121.3031
#> [10,] 243.63170 231.07525 216.36111 200.73884 185.14279 175.73159 167.7826
#>           [,15]
#>  [1,] 160.98330
#>  [2,] 192.59464
#>  [3,] 239.26773
#>  [4,] 220.33149
#>  [5,] 236.74657
#>  [6,] 236.74657
#>  [7,] 116.87544
#>  [8,]  62.51013
#>  [9,] 112.29692
#> [10,] 160.98330