farver-package {farver}	R Documentation

farver: High Performance Colour Space Manipulation

Description

The encoding of colour can be handled in many different ways, using different colour spaces. As different colour spaces have different uses, efficient conversion between these representations are important. The 'farver' package provides a set of functions that gives access to very fast colour space conversion and comparisons implemented in C++, and offers speed improvements over the 'convertColor' function in the 'grDevices' package.

Author(s)

Maintainer: Thomas Lin Pedersen thomasp85@gmail.com (ORCID)

Authors:

Berendea Nicolae (Author of the ColorSpace C++ library)

Romain François romain@purrple.cat (ORCID)

See Also

Useful links:

https://farver.data-imaginist.com

https://github.com/thomasp85/farver

Report bugs at https://github.com/thomasp85/farver/issues

[Package farver version 2.1.0 Index]